MAINT: fix fortran compiler warnings.

+++ src/fortranapex/apex.f90
@@ -8,7 +8,7 @@ module apexmodule
   real(8), parameter       :: dtor=pi / 180D0, rtod=180D0 / pi, pid2=pi / 2D0, twopi=2D0 * pi
   real(8), parameter       :: Req=6378.1370D0, eps=1.D0 / 298.257223563D0
   real(8), parameter       :: Re=Req * (1 - eps / 3D0), ecc2=eps * (2 - eps)
-  real(4), parameter       :: missing=- 9999E0
+  real(8), parameter       :: missing=- 9999E0
 
 contains
 
@@ -182,7 +182,7 @@ subroutine apex(date, igrffilein, dlat, dlon, alt, a, alat, alon, bmag, xmag, ym
   implicit none
 
 
-  real(4), intent(in)      :: date
+  real(8), intent(in)      :: date
   character(len=1000), intent(in)   :: igrffilein
   real(8), intent(in)      :: dlat, dlon, alt
   real(8)      :: a, alon, bmag, xmag, ymag, zmag, v
